requirements.txt
aiohttp==3.8.1
APScheduler==3.9.1
nest_asyncio==1.5.5
openpyxl==3.0.10
pandas==1.3.5
pycryptodome==3.15.0
pytz==2022.1
requests==2.28.0
stomper==0.4.3
websocket_client==1.4.2
websockets==10.4
google== 3.0.0
protobuf==3.19.0